test: fix lab processor tests
