--- .editorconfig
@@ -1,10 +1,0 @@
-root = true
-
-[*]
-insert_final_newline = true
-charset = utf-8
-
-[*.java]
-indent_style = space
-indent_size = 4
-# ij_continuation_indent_size = 8
