+++ src/main/java/de/unimarburg/diz/labtofhir/mapper/LoincMapper.java
@@ -5,6 +5,7 @@ import de.unimarburg.diz.labtofhir.metric.TagCounter;
 import de.unimarburg.diz.labtofhir.model.LoincMap;
 import de.unimarburg.diz.labtofhir.model.LoincMapEntry;
 import io.micrometer.core.instrument.Metrics;
+import jakarta.annotation.PostConstruct;
 import java.io.IOException;
 import org.hl7.fhir.r4.model.Coding;
 import org.hl7.fhir.r4.model.Observation;
@@ -13,8 +14,9 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.core.io.Resource;
+import org.springframework.stereotype.Service;
 
-//@Service
+@Service
 public class LoincMapper {
 
     private final FhirProperties fhirProperties;
@@ -55,12 +57,12 @@ public class LoincMapper {
         return new LoincMap().with(mappingPackage, ',');
     }
 
+    @PostConstruct
     public LoincMapper initialize() throws Exception {
         initializeMap();
         return this;
     }
 
-    //    @PostConstruct
     private void initializeMap() throws Exception {
         if (this.mappingPackage != null) {
             this.loincMap = getSwlLoincMapping(mappingPackage);
