+++ src/test/java/de/unimarburg/diz/labtofhir/processor/LabToFhirProcessorTests.java
@@ -124,5 +124,4 @@ public class LabToFhirProcessorTests {
                 .getLaboratorySystem(), "NA")).isTrue();
         }
     }
-
 }
